PycharmProjects/imdapp-master/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.16
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.9.0
django-login-required==0.2
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.5
pytz==2022.6
PyYAML==6.0
read==0.0.2
sqlparse==0.4.3
tablib==3.2.1
typing-extensions==4.4.0
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
